refactor colaculator page
