+++ src/hooks/calculators/useEmergencyFund.js
@@ -1,5 +1,6 @@
-import { getEmergencyFund } from "@/services/calculator-service";
 import { useQuery } from "@tanstack/react-query";
+
+import { getEmergencyFund } from "@/services/calculators-service";
 
 export const useEmergency = () => {
   return useQuery({
