+++ src/hooks/categories/useCategories.js
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { getCategories } from "@/services/category-service";
+import { getCategories } from "@/services/categories-service";
 
 export const useCategories = () => {
   return useQuery({
